REFACTOR mesh creation
- move mesh creation to separate file
- meshes are created by taking their neighbors
	into account
